Switch mdk::N from u32 to u8; fix Target bug for accumulated values
Cell values (1..=9) fit in u8. Accumulated arithmetic results (sums, products,
arity, depth) use Target = u32 throughout — callers that previously passed N
where Target was required (mdd.rs, tuples.rs, operation.rs) now widen via
Target::from() or explicit casts with #[allow(clippy::cast_possible_truncation)].

+++ mathdoku/src/grid.rs
@@ -402,7 +402,7 @@ mod tests {
             assert!(sol.is_solution());
             for r in 0..3 {
                 let row_sum: u32 = (0..3)
-                    .map(|c| sol.get_values(Cell::new(r, c)).unwrap().values()[0])
+                    .map(|c| u32::from(sol.get_values(Cell::new(r, c)).unwrap().values()[0]))
                     .sum();
                 assert_eq!(row_sum, 6, "row {r} should sum to 6");
             }
