+++ mathdoku/src/mdk/mod.rs
@@ -17,7 +17,7 @@ pub(crate) mod regin;
 pub mod table;
 pub mod tuples;
 
-pub type N = u32;
+pub type N = u8;
 type Target = u32;
 
 /// Errors returned by mdk operations.
